# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vantor API",
    "version": "0.8.0-one-stop"
  },
  "paths": {
...
    "/api/v1/suppliers": {
      "get": {
        "tags": [
          "suppliers"
        ],
        "summary": "List Suppliers",
        "operationId": "list_suppliers_api_v1_suppliers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Cursor"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc",
              "title": "Order"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Suppliers Api V1 Suppliers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "suppliers"
        ],
        "summary": "Create Supplier",
        "operationId": "create_supplier_api_v1_suppliers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Supplier Api V1 Suppliers Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/suppliers/{supplier_id}": {
      "get": {
        "tags": [
          "suppliers"
        ],
        "summary": "Get Supplier",
        "operationId": "get_supplier_api_v1_suppliers__supplier_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Supplier Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Supplier Api V1 Suppliers  Supplier Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "suppliers"
        ],
        "summary": "Update Supplier",
        "operationId": "update_supplier_api_v1_suppliers__supplier_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Supplier Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Update Supplier Api V1 Suppliers  Supplier Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/suppliers/{supplier_id}/contacts": {
      "get": {
        "tags": [
          "suppliers"
        ],
        "summary": "List Contacts",
        "operationId": "list_contacts_api_v1_suppliers__supplier_id__contacts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Supplier Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Contacts Api V1 Suppliers  Supplier Id  Contacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "suppliers"
        ],
        "summary": "Add Contact",
        "operationId": "add_contact_api_v1_suppliers__supplier_id__contacts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Supplier Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Add Contact Api V1 Suppliers  Supplier Id  Contacts Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/suppliers/{supplier_id}/certifications/{cert_id}/verify": {
      "post": {
        "tags": [
          "suppliers"
        ],
        "summary": "Verify Cert",
        "operationId": "verify_cert_api_v1_suppliers__supplier_id__certifications__cert_id__verify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Supplier Id"
            }
          },
          {
            "name": "cert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cert Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Verify Cert Api V1 Suppliers  Supplier Id  Certifications  Cert Id  Verify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContactIn": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "default": ""
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "default": ""
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "full_name"
        ],
        "title": "ContactIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SupplierIn": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 2,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 300,
            "minLength": 2,
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "draft"
          },
          "country": {
            "type": "string",
            "title": "Country",
            "default": ""
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": ""
          },
          "category_id": {
            "type": "string",
            "title": "Category Id",
            "default": ""
          },
          "payment_terms": {
            "type": "string",
            "title": "Payment Terms",
            "default": ""
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "SupplierIn"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}